Minor edits in the database table

+++ migrations/2018_07_11_140701_create_error_notifications_table.php
@@ -9,7 +9,7 @@ class CreateErrorNotificationsTable extends Migration
     public function up()
     {
         Schema::create('error_notifications', function (Blueprint $table) {
-            $table->increments('id');
+            $table->unsignedBigInteger('id');
 
             $table->string('parent');
             $table->longText('exception');
